+++ source/MIBRR/R/02_EstimationMethods.R
@@ -1,7 +1,7 @@
 ### Title:    Optimization and Gibbs Sampling Methods for MIBRR
 ### Author:   Kyle M. Lang
 ### Created:  2017-SEP-30
-### Modified: 2018-JUN-07
+### Modified: 2018-JUN-13
 ### Notes:    This file will add optimization and Gibbs sampling methods to the
 ###           MibrrFit class.
 
@@ -55,8 +55,9 @@ MibrrFit$methods(
                               betaStarts  = betaStarts,
                               burnSams    = sampleSizes[[phase]][1],
                               totalSams   = sum(sampleSizes[[phase]]),
+                              penType     = penalty,
+                              ridge       = ridge,
                               verbose     = verbose,
-                              doBl        = doBl,
                               fullBayes   = !doMcem,
                               adaptScales = adaptScales,
                               noMiss      = all(missCounts == 0),
@@ -171,7 +172,7 @@ MibrrFit$methods(
                  "Optimize the BEN or BL penalty parameters"
                  
                  ## Use simple update rule and return early when doing BL:
-                 if(doBl) {
+                 if(penalty == 2) {
                      lapply(1 : nTargets, .self$updateBlLambda)
                  }
                  else {
@@ -194,7 +195,7 @@ MibrrFit$methods(
                      
                      if(optTraceLevel == 0) sink()
                      options(warn = 0)
-                 }# CLOSE if(doBl); else
+                 }# CLOSE if(penalty == 2); else
                  
                  for(j in 1 : nTargets) {
                      lambdaHistory[[j]][iter, ] <<- lambdaMat[j, ]