+++ CHANGELOG.md
@@ -14,7 +14,13 @@ NOTE:
 - On 2017-11-06 the package name was changed from `mibrr` to `MIBRR`, so the 
   version number was reset to 0.0.0.9000, as well
 
-## 0.0.0.9006 - 2018-06-07 - ACTIVE
+## 0.0.0.9007 - 2018-06-13 - ACTIVE
+
+### Added
+- A new exported function, `vanilla`, that implements a basic MI without 
+  using any shrinkage priors.
+
+## 0.0.0.9006 - 2018-06-07
 
 ### Fixed
 - Bug causing crashes when initially imputing incomplete auxiliary variables